feat: 결과를 계산하는 클래스 추가
